Tool

+++ Tool/Z.Infra.StatItemListSourceGen/AvalonGen.cs
@@ -10,7 +10,7 @@ public class AvalonGen : AvalonSourceGen
         return true;
     }
 
-    public override int Execute()
+    public override long Execute()
     {
         this.OutputFilePath = this.GetOutputFilePath();
         return base.Execute();
